+++ source/rockhopper/core/package.d
@@ -5,3 +5,4 @@ public import rockhopper.core.suspends;
 public import rockhopper.core.reactor;
 public import rockhopper.core.threading;
 public import rockhopper.core.llevents;
+public import rockhopper.core.uda;